Well, this commit adds var arguments (hopefully).  It was a tough decision on how to implement them.
I had a few choices:  One was to make everything being stored as a variable a pointer, which would be most true to the idea of pascal, but the way java works, that would mean there would be an extra object for every variable, on top of the Integer or Double wrapper type, as they are unmodifiable classes.  This would (I think) cause a huge memory requirement.

The other Idea was to pass a reference to the variable in the form of the calling functions hashmap and the variable name, but that also would cause problems when variables are repetitively passed by reference, and it means I would have to add a special check to all my variable modification code to check if it is a reference or actual value.  This would (I think) make things slower.

The idea I went with was to create a new class called pointer, which was basically similar to the first idea, except it is only used for the transfer of variables to a function.  They are wrapped before, stored in the function on the stack, and then after the function is done executing, they are updated in their values.  Then the abstract function call re-updates the variables in the calling function using a variable_set instruction.  This seems like it will be slower, but it will only affect things when var args are used, and it does not require processing every time a variable is acessed, only when entering and exiting a function.  It also requires more memory for storing the pointer objects during the execution of the function.   It is also easier to implement, and has the nice side effect that plugins only need to ask for a pointer type argument to get a var argument (I realize now, that I will need to update the type checking to make up for the fact that anything  var args will have the class pointer.  But what was nice is that it didn't break the abstraction too badly.

git-svn-id: https://pascalinterpreterinjava.googlecode.com/svn/trunk@56 fb5b4802-194b-0410-9e11-5134f6b347ec