I tried to work out a minimized client screen reader, but I couldn't get it to work after a while.  I still have the code, but I know it won't work this way.
Also, I started adding the real SCAR emulation! I have another project SCARlib that adds native stuff that java can't do.  It only works with X right now, but it actually should be easier to write for windows.  It uses some code from xwininfo.  I think I will upload it as another project when it gets to a stable point.  It isn't that much because the Robot class provides most of the functionality I need.

git-svn-id: https://pascalinterpreterinjava.googlecode.com/svn/trunk@104 fb5b4802-194b-0410-9e11-5134f6b347ec